requirements.txt
# Core web scraping framework
scrapy>=2.5.0

# Database connectivity
psycopg2-binary>=2.9.0

# HTTP requests
requests>=2.28.0

# Twitter API integration
tweepy>=4.14.0

# LangChain and Groq LLM integration
langchain>=0.1.0
langchain-groq>=0.1.0



# Additional dependencies that might be needed
# URL parsing (usually included with Python)
urllib3>=1.26.0

# Date and time handling (usually included with Python)
python-dateutil>=2.8.0

# Optional: For better logging and debugging
colorlog>=6.7.0

# Optional: For configuration management
python-dotenv>=1.0.0